requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
attrs
bcrypt
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
et_xmlfile
filelock
frozenlist
fsspec
gitdb
GitPython
huggingface-hub
idna
inquirerpy
Jinja2
jsonschema
jsonschema-specifications
lml
markdown-it-py
MarkupSafe
mdurl
multidict
narwhals
numpy
opencv-python-headless
openpyxl
packaging
pandas
pfzy
pillow
prompt_toolkit
propcache
protobuf
psycopg2-binary
pyarrow
pydeck
pyexcel
pyexcel-io
pyexcel-xls
Pygments
python-dateutil
python-dotenv
pytz
PyYAML
referencing
requests
rich
rpds-py
six
smmap
soupsieve
SQLAlchemy
streamlit
tenacity
texttable
toml
tornado
tqdm
typing_extensions
tzdata
urllib3
wcwidth
xlrd
xlwt
yarl